+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- The directory of GTlab's config files has been changed on Linux to ~/.config/DLR/GTlab - #327
 
 ## [2.0.0-pr-alpha6] - 2023-02-03
 ### Changed